requirements.txt
streamlit==1.51.0
pandas
numpy
scikit-learn
joblib
Pillow
